+++ src/platformAccessory.ts
@@ -38,13 +38,13 @@ class TemperatureSensor {
   private getCurrentTemperature(): number {
 
     try {
-      const value_key: string = this.accessory.context.device.value_key;
+      const value_key: string = this.accessory.context.device.temperature_properties.value_key;
       const temperature: number = parseFloat(this.platform.observation_data[value_key]);
       if (temperature > 100.00) {
-        this.platform.log.debug(`WeatherFlow Tempest is reporting temperatures exceeding 100c: ${temperature}c`);
+        this.platform.log.debug(`WeatherFlow Tempest is reporting temperatures exceeding 100C: ${temperature}C`);
         return 100;
       } else if (temperature < -271.00) {
-        this.platform.log.debug(`WeatherFlow Tempest is reporting temperatures less than -271c: ${temperature}c`);
+        this.platform.log.debug(`WeatherFlow Tempest is reporting temperatures less than -271C: ${temperature}C`);
         return -271.00;
       } else {
         return temperature;
@@ -98,7 +98,7 @@ class LightSensor {
   private getCurrentLux(): number {
 
     try {
-      const value_key: string = this.accessory.context.device.value_key;
+      const value_key: string = this.accessory.context.device.light_properties.value_key;
       const lux: number = parseFloat(this.platform.observation_data[value_key]);
       if (lux < 0.0001) {
         this.platform.log.debug(`WeatherFlow Tempest is reporting lux less than 0.0001: ${lux}`);
@@ -158,7 +158,7 @@ class HumiditySensor {
   private getCurrentRelativeHumidity(): number {
 
     try {
-      const value_key: string = this.accessory.context.device.value_key;
+      const value_key: string = this.accessory.context.device.humidity_properties.value_key;
       const relative_humidity: number = parseInt(this.platform.observation_data[value_key]);
       if (relative_humidity > 100) {
         this.platform.log.debug(`WeatherFlow Tempest is reporting relative humidity exceeding 100%: ${relative_humidity}%`);
@@ -205,7 +205,9 @@ class MotionSensor {
       .onGet(this.handleMotionDetectedGet.bind(this));
 
     // Set initial value
-    this.service.getCharacteristic(this.platform.Characteristic.MotionDetected).updateValue(this.isMotionDetected());
+    this.service.getCharacteristic(
+      this.platform.Characteristic.MotionDetected).updateValue(this.isMotionDetected(),
+    );
 
     // Update value based on user defined global interval
     const interval = (this.platform.config.interval as number || 10) * 1000;
@@ -215,11 +217,18 @@ class MotionSensor {
 
   }
 
-  private getValue(): number {
+  private getMotionSensorValue(): number {
 
     try {
-      const value_key: string = this.accessory.context.device.value_key;
-      const speed: number = parseInt(this.platform.observation_data[value_key]);
+      const value_key: string = this.accessory.context.device.motion_properties.value_key;
+      let speed: number = parseFloat(this.platform.observation_data[value_key]);
+
+      if (this.platform.config.units === 'Metric') {
+        speed = Math.round(speed);
+      } else {
+        speed = Math.round(speed * 2.236936); // convert m/s to mph
+      }
+
       if (speed < 0) {
         this.platform.log.debug(`WeatherFlow Tempest Motion Sensor is reporting less than 0: ${speed}`);
         return 0;
@@ -235,15 +244,15 @@ class MotionSensor {
 
   private isMotionDetected(): boolean {
 
-    const current_value = this.getValue();
-    let motion_trigger_value = 1;
+    const current_value = this.getMotionSensorValue();
+    let trigger_value = 1;
     try {
-      motion_trigger_value = this.accessory.context.device.additional_properties.motion_trigger_value;
+      trigger_value = this.accessory.context.device.motion_properties.trigger_value;
     } catch(exception) {
       this.platform.log.error(exception as string);
       this.platform.log.warn('Defaulting to 1 as motion trigger value.');
     }
-    return current_value >= motion_trigger_value;
+    return current_value >= trigger_value;
 
   }
 
@@ -291,8 +300,9 @@ class Fan {
   private getCurrentWindSpeed(): number {
 
     try {
-      const value_key: string = this.accessory.context.device.value_key;
-      const speed: number = parseInt(this.platform.observation_data[value_key]);
+      const value_key: string = this.accessory.context.device.fan_properties.value_key;
+      let speed: number = parseFloat(this.platform.observation_data[value_key]);
+      speed = Math.round(speed * 2.236936); // convert m/s to mph and round as fan % is integer value
       if (speed > 100) {
         this.platform.log.debug(`WeatherFlow Tempest is reporting wind speed exceeding 100mph: ${speed}mph`);
         return 100;
@@ -321,6 +331,289 @@ class Fan {
 
 }
 
+class OccupancySensor {
+  private service: Service;
+
+  constructor(
+    private readonly platform: WeatherFlowTempestPlatform,
+    private readonly accessory: PlatformAccessory,
+  ) {
+
+    // Restore or create Occupancy Sensor
+    this.service = this.accessory.getService(this.platform.Service.OccupancySensor) ||
+    this.accessory.addService(this.platform.Service.OccupancySensor);
+
+    // Create handlers for required characteristics
+    this.service.getCharacteristic(this.platform.Characteristic.OccupancyDetected)
+      .onGet(this.handleOccupancyDetectedGet.bind(this));
+
+    // Set initial state of Occupancy Sensor and sensor value
+    const sensorName = this.accessory.context.device.name;
+    const [sensorValue, sensorUnits, sensorTrip] = this.getOccupancySensorValue();
+
+    this.service.getCharacteristic(
+      this.platform.Characteristic.Name).updateValue(`${sensorName}: ${sensorValue} ${sensorUnits}`,
+    );
+
+    this.service.getCharacteristic(
+      this.platform.Characteristic.OccupancyDetected).updateValue((sensorValue >= sensorTrip),
+    );
+
+    // Update occupancy sensor state and name based on user defined global interval
+    const interval = (this.platform.config.interval as number || 10) * 1000;
+
+    setInterval( () => {
+      const sensorName = this.accessory.context.device.name;
+      const [sensorValue, sensorUnits, sensorTrip] = this.getOccupancySensorValue();
+
+      this.service.getCharacteristic(
+        this.platform.Characteristic.Name).updateValue(`${sensorName}: ${sensorValue} ${sensorUnits}`,
+      );
+
+      this.service.getCharacteristic(
+        this.platform.Characteristic.OccupancyDetected).updateValue((sensorValue >= sensorTrip),
+      );
+
+    }, interval);
+
+  }
+
+  private getOccupancySensorValue(): [value: number, units: string, trip_level: number] {
+
+    try {
+      const value_key: string = this.accessory.context.device.occupancy_properties.value_key;
+      let trigger_value = this.accessory.context.device.occupancy_properties.trigger_value;
+      let value: number = parseFloat(this.platform.observation_data[value_key]);
+      let units = '';
+
+      // check that trip_level is not less than 0
+      if (trigger_value < 0) {
+        trigger_value = 0;
+      }
+
+      switch (value_key) {
+        case 'barometric_pressure':
+          if (this.platform.config.units === 'Metric') {
+            value = Math.round(value * 1000) / 1000; // 3 decimal places
+            units = 'mb';
+          } else {
+            value = Math.round(value / 33.8638 * 1000) / 1000; // convert from mb to inHg
+            units = 'inHg';
+          }
+          break;
+        case 'precip': //
+          if (this.platform.config.units === 'Metric') {
+            value = Math.round(value * 100) / 100; // 2 decimal places
+            units = 'mm/min';
+          } else {
+            value = Math.round(value * 2.36 * 100) / 100; // convert mm/min to in/hr
+            units = 'in/hr';
+          }
+          break;
+        case 'precip_accum_local_day': // convert mm to in
+          if (this.platform.config.units === 'Metric') {
+            value = Math.round(value * 100) / 100; // 2 decimal places
+            units = 'mm';
+          } else {
+            value = Math.round(value / 25.4 * 100) / 100; // convert mm to in
+            units = 'in';
+          }
+          break;
+        case 'solar_radiation': // no conversion needed
+          units = 'W/m\xB2';
+          break;
+        case 'uv': // no conversion needed
+          value = Math.round(value * 10) / 10; // 1 decimal place
+          units = ' ';
+          break;
+        case 'wind_direction': // convert from degrees to cardinal (N, S, E, W)
+          // eslint-disable-next-line no-case-declarations
+          const cat = Math.round(value % 360 / 22.5);
+          switch (cat) {
+            case 0:
+              units = '\xB0 N';
+              break;
+            case 1:
+              units = '\xB0 NNE';
+              break;
+            case 2:
+              units = '\xB0 NE';
+              break;
+            case 3:
+              units = '\xB0 ENE';
+              break;
+            case 4:
+              units = '\xB0 E';
+              break;
+            case 5:
+              units = '\xB0 ESE';
+              break;
+            case 6:
+              units = '\xB0 SE';
+              break;
+            case 7:
+              units = '\xB0 SSE';
+              break;
+            case 8:
+              units = '\xB0 S';
+              break;
+            case 9:
+              units = '\xB0 SSW';
+              break;
+            case 10:
+              units = '\xB0 SW';
+              break;
+            case 11:
+              units = '\xB0 WSW';
+              break;
+            case 12:
+              units = '\xB0 W';
+              break;
+            case 13:
+              units = '\xB0 WNW';
+              break;
+            case 14:
+              units = '\xB0 NW';
+              break;
+            case 15:
+              units = '\xB0 NNW';
+              break;
+            case 16:
+              units = '\xB0 N';
+              break;
+            default:
+              units = ' Variable';
+          }
+          break;
+        default:
+          break;
+      }
+
+      if (value < 0) {
+        value = 0;
+        this.platform.log.debug(`WeatherFlow Tempest ${value_key} is reporting less than 0: ${value}`);
+        return [value, units, trigger_value];
+
+      } else {
+        this.platform.log.debug(`WeatherFlow Tempest ${value_key}: ${value} ${units}, trip_level: ${trigger_value}`);
+        return [value, units, trigger_value];
+      }
+
+    } catch(exception) {
+      this.platform.log.error(exception as string);
+      return [0, '', 1000];
+    }
+
+  }
+
+  private isOccupancyDetected(): boolean {
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const [sensorValue, sensorUnits, sensorTrip] = this.getOccupancySensorValue();
+    return sensorValue >= sensorTrip;
+
+  }
+
+  /**
+   * Handle requests to get the current value of the "Motion Detected" characteristic
+   */
+  private handleOccupancyDetectedGet(): boolean {
+
+    this.platform.log.debug('Triggered GET MotionDetected');
+    return this.isOccupancyDetected();
+
+  }
+
+}
+
+class BatterySensor {
+  private service: Service;
+
+  constructor(
+    private readonly platform: WeatherFlowTempestPlatform,
+    private readonly accessory: PlatformAccessory,
+  ) {
+
+    // Get or Add Service to Accessory
+    this.service = this.accessory.getService(this.platform.Service.Battery) ||
+    this.accessory.addService(this.platform.Service.Battery, 'Tempest Battery');
+
+    // Create handlers for required characteristics
+    this.service.getCharacteristic(this.platform.Characteristic.BatteryLevel)
+      .onGet(this.handleCurrentBatteryLevelGet.bind(this));
+
+    // Set Current Battery Level
+    this.service.getCharacteristic(
+      this.platform.Characteristic.BatteryLevel).updateValue(this.getCurrentBatteryLevel(),
+    );
+
+    // Update value based on user defined global interval
+    const interval = (this.platform.config.interval as number || 10) * 1000;
+    setInterval( () => {
+      this.service.getCharacteristic(
+        this.platform.Characteristic.BatteryLevel).updateValue(this.getCurrentBatteryLevel(),
+      );
+    }, interval);
+
+  }
+
+  /**
+   * Get the current battery level from the global observation object.
+   */
+  private getCurrentBatteryLevel(): number {
+
+    try {
+      const batteryLevel: number = this.platform.tempest_battery_level;
+      if (batteryLevel > 100) {
+        this.platform.log.debug(`WeatherFlow Tempest is reporting battery level exceeding 100%: ${batteryLevel}%`);
+        return 100;
+      } else if (batteryLevel < 0) {
+        this.platform.log.debug(`WeatherFlow Tempest is reporting battery level less than 0%: ${batteryLevel}%`);
+        return 0;
+      } else {
+        return batteryLevel;
+      }
+    } catch(exception) {
+      this.platform.log.error(exception as string);
+      return 0;
+    }
+
+  }
+
+  /**
+   * Handle requests to get the current value of the "Current Battery Level" characteristic
+   */
+  private handleCurrentBatteryLevelGet(): number {
+
+    this.platform.log.debug('Triggered GET CurrentBatteryLevel');
+    const batteryLevel: number = this.getCurrentBatteryLevel();
+    return batteryLevel;
+  }
+
+}
+
+/**
+ * Initialize Tempest Platform (only need to do once)
+ */
+export class InitWeatherFlowTempestPlatform {
+
+  constructor(
+    private readonly platform: WeatherFlowTempestPlatform,
+    private readonly accessory: PlatformAccessory,
+  ) {
+
+    // Configure Accessory
+    this.accessory.getService(this.platform.Service.AccessoryInformation)!
+      .setCharacteristic(this.platform.Characteristic.Manufacturer, 'WeatherFlow')
+      .setCharacteristic(this.platform.Characteristic.Model, 'Tempest')
+      .setCharacteristic(this.platform.Characteristic.SerialNumber, `${this.platform.config.station_id}`);
+
+    // Add battery sensor
+    new BatterySensor(this.platform, this.accessory);
+
+  }
+}
+
 /**
  * Platform Accessory
  */
@@ -335,12 +628,16 @@ export class WeatherFlowTempestPlatformAccessory {
     this.accessory.getService(this.platform.Service.AccessoryInformation)!
       .setCharacteristic(this.platform.Characteristic.Manufacturer, 'WeatherFlow')
       .setCharacteristic(this.platform.Characteristic.Model, `Tempest - ${this.accessory.context.device.name}`)
-      .setCharacteristic(this.platform.Characteristic.SerialNumber, '000');
+      .setCharacteristic(this.platform.Characteristic.SerialNumber, `${this.platform.config.station_id}`);
 
-    // ["Temperature Sensor", "Light Sensor", "Humidity Sensor", "Fan", "Motion Sensor"]
     switch (this.accessory.context.device.sensor_type) {
       case 'Temperature Sensor':
         new TemperatureSensor(this.platform, this.accessory);
+
+        // Add Battery to default Temperature air_temperature sensor
+        if (this.accessory.context.device.temperature_properties.value_key === 'air_temperature') {
+          new BatterySensor(this.platform, this.accessory);
+        }
         break;
       case 'Light Sensor':
         new LightSensor(this.platform, this.accessory);
@@ -353,6 +650,9 @@ export class WeatherFlowTempestPlatformAccessory {
         break;
       case 'Fan':
         new Fan(this.platform, this.accessory);
+        break;
+      case 'Occupancy Sensor':
+        new OccupancySensor(this.platform, this.accessory);
         break;
     }