+++ src/tempestApi.ts
@@ -2,15 +2,28 @@ import { Logger } from 'homebridge';
 import axios, { AxiosResponse } from 'axios';
 
 export interface Observation {
-  air_temperature: number;
+  // temperature sensors
+  air_temperature: number;         // C, displayed according to Homebridge and HomeKit C/F settings
   feels_like: number;
   wind_chill: number;
   dew_point: number;
-  relative_humidity: number;
-  wind_avg: number;
-  wind_gust: number;
-  wind_direction: number;
-  brightness: number;
+
+  // humidity sensor
+  relative_humidity: number;       // %
+
+  // fan and motion sensor
+  wind_avg: number;                // m/s, used for Fan speed %
+  wind_gust: number;               // m/s, used for motion sensor
+
+  // occupancy sensors
+  barometric_pressure: number;     // mb
+  precip: number;                  // mm/min (minute sampling)
+  precip_accum_local_day: number;  // mm
+  wind_direction: number;          // degrees
+  solar_radiation: number;         // W/m^2
+  uv: number;                      // Index
+
+  brightness: number;              // Lux
 }
 
 export class TempestApi {
@@ -19,6 +32,8 @@ export class TempestApi {
   private token: string;
   private station_id: string;
   private data: object | undefined;
+  private tempest_device_id: number;
+  private tempest_battery_level: number;
   private readonly max_retries: number = 30;
 
   constructor(token: string, station_id: string, log: Logger) {
@@ -27,6 +42,8 @@ export class TempestApi {
     this.token = token;
     this.station_id = station_id;
     this.data = undefined; // last sample of Observation data
+    this.tempest_device_id = 0;
+    this.tempest_battery_level = 0;
 
     this.log.info('TempestApi initialized.');
 
@@ -98,6 +115,52 @@ export class TempestApi {
       this.data = response.data['obs'][0];
       return this.data;
     }
+
+  }
+
+  public async getTempestBatteryLevel(device_id): Promise<number> {
+
+    const url = `https://swd.weatherflow.com/swd/rest/observations/device/${device_id}`;
+    const options = {
+      headers: {
+        'Authorization': `Bearer ${this.token}`,
+      },
+      validateStatus: (status: number) => status < 500, // Resolve only if the status code is less than 500
+    };
+
+    await axios.get(url, options) // assumes single Tempest station
+      .then(response => {
+        const tempest_battery_voltage = response.data.obs[0][16];
+        this.tempest_battery_level = Math.round((tempest_battery_voltage - 1.8) * 100); // 2.80V = 100%, 1.80V = 0%
+      })
+
+      .catch(exception => {
+        this.log.debug(`[WeatherFlow] ${exception}`);
+      });
+
+    return this.tempest_battery_level;
+
+  }
+
+  public async getTempestDeviceId(): Promise<number> {
+
+    const url = `https://swd.weatherflow.com/swd/rest/stations/${this.station_id}`;
+    const options = {
+      headers: {
+        'Authorization': `Bearer ${this.token}`,
+      },
+      validateStatus: (status: number) => status < 500, // Resolve only if the status code is less than 500
+    };
+
+    await axios.get(url, options) // assumes single hub with single Tempest station
+      .then(response => {
+        this.tempest_device_id = response.data.stations[0].devices[1].device_id;
+      })
+
+      .catch(exception => {
+        this.log.debug(`[WeatherFlow] ${exception}`);
+      });
+    return this.tempest_device_id;
 
   }